Revert "Merge pull request #3 from a5021/feature/non-blocking-device-search"
This reverts commit dcdc23e78912a31376acb117927a0ca3e2131b73, reversing
changes made to f152deb48fea102b292444c5459044b87cb161aa.
