+++ inc/ds18b20.h
@@ -7,23 +7,15 @@
  * 
  * Key features:
  * - Pure bare-metal, register-level programming
- * - No interrupts; no software delays or busy-waits in the measurement path
+ * - No interrupts, no software delays, no busy-waits
  * - Hardware timer-based timing with DMA for data capture
  * - Non-blocking state machine architecture
  * - Weak function callbacks for customization
- * - Built-in non-blocking device search (ds18b20_search_*) for multi-sensor
- *   buses, with zero busy-waits
- * 
- * All low-level 1-Wire bus operations and the Search ROM state machine live
- * inside the driver. The public interface is intentionally small: start and
- * poll the driver, select a device, and (optionally) run a bus scan.
  * 
  * Usage:
  * 1. Call ds18b20_init() once at startup
  * 2. Call ds18b20_poll() repeatedly from main loop
- * 3. (Optional) Run the non-blocking device search (ds18b20_search_*) before
- *    starting poll(); the search helper hands back to poll() automatically
- * 4. Implement weak callbacks ds18b20_busy() and ds18b20_complete()
+ * 3. Implement weak callbacks ds18b20_busy() and ds18b20_complete()
  *    to handle status indication and temperature results
  */
 
@@ -50,58 +42,6 @@ extern "C" {
 #define DS18B20_TEMP_ERROR_CRC_FAIL (INT16_MIN + 2) /**< CRC checksum validation failed */
 
 /**
- * @brief DS18B20 1-Wire family code (LSB of the 64-bit ROM address)
- * @note Other 1-Wire devices on the bus use different family codes and are
- *       skipped by the device search.
- */
-#define DS18B20_FAMILY_CODE 0x28
-
-/**
- * @brief Number of bytes in a device ROM address
- */
-#define DS18B20_ROM_BYTES 8
-
-/**
- * @}
- */
-
-/**
- * @defgroup DS18B20_Search DS18B20 Non-Blocking Device Search
- * @brief Maxim Search ROM (0xF0) state machine, driven from the main loop
- *        like the measurement state machine. It filters by family code,
- *        validates the CRC and reports each found DS18B20 via a callback.
- *        When the search finishes it hands the timer back to ds18b20_poll().
- * @{
- */
-
-/**
- * @brief Callback invoked for every DS18B20 found by the search
- * @param[in] rom Pointer to the 8-byte ROM address (LSB first)
- * @return 0 to continue the search, non-zero to stop
- * @note The pointer is only valid for the duration of the callback.
- */
-typedef uint8_t (*ds18b20_search_sink_t)(const uint8_t* rom);
-
-/**
- * @brief Start a non-blocking device search
- * @param[in] sink Callback invoked per found DS18B20 device (may be NULL)
- * @param[in] max_devices Maximum number of devices to report (0 aborts)
- */
-void ds18b20_search_start(ds18b20_search_sink_t sink, uint8_t max_devices);
-
-/**
- * @brief Advance the non-blocking device search by one hardware operation
- * @return 1 when the search is finished, 0 while still running
- */
-uint8_t ds18b20_search_poll(void);
-
-/**
- * @brief Number of DS18B20 devices found (valid once the search finished)
- * @return Count of found devices
- */
-uint8_t ds18b20_search_count(void);
-
-/**
  * @}
  */
 
@@ -118,24 +58,12 @@ void ds18b20_init(void);
 /**
  * @brief Advance the state machine (non-blocking)
  * @note Call periodically from main loop
- *
+ * 
  * This function implements the core non-blocking state machine that manages
  * the 1-Wire communication protocol with the DS18B20 sensor. It uses hardware
  * timer and DMA to handle timing-critical operations without software delays.
  */
 void ds18b20_poll(void);
-
-/**
- * @brief Select which DS18B20 device to measure by its ROM address
- * @param[in] rom Pointer to the 8-byte ROM address (LSB first), or NULL to
- *                return to Skip ROM (broadcast) addressing
- * @note With a non-NULL address, the state machine sends Match ROM (0x55)
- *       plus the device address before each command, so only that device
- *       responds. Pass NULL to keep the legacy single-sensor Skip ROM
- *       behaviour. To measure a specific device, pass its ROM address
- *       (e.g., from a bus search) to ds18b20_select().
- */
-void ds18b20_select(const uint8_t* rom);
 
 /**
  * @brief Busy indicator callback (weak)
